Create 03_annotations_DGE_Dotplots_and_final_gene_list.R
After inspection of plots and gene lists calculated with FindAllMarkers for different resulutions, an annotation of cluster can be decided. A final gene list can be computed with merged clusters, then differentially expressed genes can be calculated in each cluster between HV and CD+IBS for example - an other 1 vs 1 interactions. Volcano plots are a great way of visualising DGE, and dotplots are a great way to visualise how a specific gene is expressed in a certain subtype